refactor: improve score handling and date formatting
- Change score field type from integer to decimal for precision
- Standardize date format across score-related operations
- Add missing model imports for better code organization

+++ database/migrations/2024_01_15_000005_create_scores_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('scores', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
-            $table->integer('skor');
+            $table->decimal('skor', 5, 2);
             $table->integer('total_soal');
             $table->integer('benar');
             $table->integer('salah');
